Let TemplateController have access to the character_split method by moving it to a concern
